Move C-expression globals to the macro glue code
Removes the last dependency of `hs-bindgen:internal` on `c-expr-runtime`
by moving the `CExpr`-specific globals from `Backend.Global` into
`HsBindgen.Macro`.
