+++ hs-bindgen/app/HsBindgen/Cli/GenTests.hs
@@ -21,7 +21,6 @@ import HsBindgen.App
 import HsBindgen.Config
 import HsBindgen.Config.Internal (BindgenConfig)
 import HsBindgen.Frontend.RootHeader
-import HsBindgen.IO
 import HsBindgen.Macro
 
 {-------------------------------------------------------------------------------
