+++ hs-bindgen/app/HsBindgen/Cli/Info/UseDeclGraph.hs
@@ -22,7 +22,6 @@ import HsBindgen.Config
 import HsBindgen.Config.Internal (BindgenConfig)
 import HsBindgen.Frontend.RootHeader
 import HsBindgen.Imports
-import HsBindgen.IO
 import HsBindgen.Macro
 
 {-------------------------------------------------------------------------------
