+++ hs-bindgen/app/HsBindgen/Cli/Internal/Frontend.hs
@@ -24,7 +24,6 @@ import HsBindgen.Config
 import HsBindgen.Config.Internal
 import HsBindgen.Frontend.RootHeader
 import HsBindgen.Imports
-import HsBindgen.IO
 import HsBindgen.Macro
 
 {-------------------------------------------------------------------------------
