+++ hs-bindgen/app/HsBindgen/Cli/Preprocess.hs
@@ -25,7 +25,6 @@ import HsBindgen.Config
 import HsBindgen.Config.Internal
 import HsBindgen.Frontend.RootHeader
 import HsBindgen.Imports
-import HsBindgen.IO
 import HsBindgen.Macro
 
 {-------------------------------------------------------------------------------
