+++ hs-bindgen/src/HsBindgen/Macro/Translation/Common.hs
@@ -1,0 +1,3 @@
+module HsBindgen.Macro.Translation.Common (
+
+  ) where
